pyramidProject/requirements.txt
arrow==1.2.3
bidict==0.22.1
binaryornot==0.4.4
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
hupper==1.12
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.3
netifaces==0.10.6
PasteDeploy==3.0.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
pyramid==2.0.1
python-dateutil==2.8.2
python-engineio==4.4.1
python-slugify==8.0.1
python-socketio==5.8.0
PyYAML==6.0
requests==2.31.0
six==1.16.0
text-unidecode==1.3
translationstring==1.4
urllib3==2.0.3
venusian==3.0.0
waitress==2.1.2
WebOb==1.8.7
zope.deprecation==5.0
zope.event==5.0
zope.interface==6.0
